move the test register_system_cached_on_deferred_world to the correct location
